requirements.txt
flask>=3.0,<4.0
flask-sqlalchemy>=3.1,<4.0
pyyaml>=6.0,<7.0
